pre api

+++ system/templates/deployment/global/build_scripts/package_installer.sh
@@ -95,7 +95,12 @@ if test -d  $destination
  then
     # extract into
     echo "cp -rp ./$path_to_extracted/. $destination"
- 	cp -rp "./$path_to_extracted/." $destination
+    if test -f  ./$path_to_extracted 
+     then
+     	cp -rp "./$path_to_extracted" $destination
+     else
+ 		cp -rp "./$path_to_extracted/." $destination
+ 	 fi
  else
  echo "./$path_to_extracted $destination"
  	mv "./$path_to_extracted" $destination
